# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TRUE CONCEPT Learning Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: getSubjects
      tags: [subjects]
      summary: List all subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{subjectId}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    put:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject (admin only)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject (admin only)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /progress/mcq-score:
    post:
      operationId: saveMcqScore
      tags: [progress]
      summary: Save MCQ quiz score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveMcqScoreBody"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
  /progress/mark-chapter:
    post:
      operationId: markChapterVisited
      tags: [progress]
      summary: Mark chapter as visited
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkChapterBody"
      responses:
        "200":
          description: Marked
...
components:
  schemas:
...
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        classLevels:
          type: array
          items:
            type: string
        color:
          type: string
        chapterCount:
          type: integer
      required:
        - id
        - name
        - description
        - icon
        - classLevels
        - color
    CreateSubjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        classLevels:
          type: array
          items:
            type: string
        color:
          type: string
      required:
        - name
        - description
        - icon
        - classLevels
        - color
...
    Progress:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        chapterId:
          type: integer
        chapterTitle:
          type: string
        subjectName:
          type: string
        mcqScore:
          type: integer
        mcqTotal:
          type: integer
        visited:
          type: boolean
        lastAccessedAt:
          type: string
      required:
        - id
        - userId
        - chapterId
        - chapterTitle
        - subjectName
        - visited
    SaveMcqScoreBody:
      type: object
      properties:
        chapterId:
          type: integer
        score:
          type: integer
        total:
          type: integer
      required:
        - chapterId
        - score
        - total
    MarkChapterBody:
      type: object
      properties:
        chapterId:
          type: integer
      required:
        - chapterId
...