# lib/api-zod/src/generated/types/markChapterBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TRUE CONCEPT Learning Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface MarkChapterBody {
  chapterId: number;
}